+++ packages/runtime/src/contentSources.ts
@@ -27,14 +27,31 @@ export type ContentSourceDependenciesValueFromRef<
     >,
 > = {
     // TODO: extend to support other types once ComputedContentDependencyRef becomes a union
-    [K in keyof Dependencies]: ComputedContentDependencyValue;
+    [K in keyof Dependencies]: Extract<
+        ComputedContentDependencyValue,
+        { ref: Dependencies[K]['ref'] }
+    >;
+};
+
+export type ContentSourceInput<
+    Props extends PlainObject = {},
+    Dependencies extends Record<
+        string,
+        {
+            ref: ComputedContentDependencyRef;
+        }
+    > = {},
+> = {
+    props: Props;
+    dependencies: ContentSourceDependenciesValueFromRef<Dependencies>;
 };
 
 /**
  * Create a content source. The result should be bind to the integration using `contentSources`.
  */
 export function createContentSource<
-    Props extends PlainObject = {},
+    GetRevisionProps extends PlainObject = {},
+    GetPageDocumentProps extends PlainObject = {},
     Dependencies extends Record<
         string,
         {
@@ -52,12 +69,7 @@ export function createContentSource<
      * Callback to generate the pages.
      */
     getRevision: RuntimeCallback<
-        [
-            {
-                props: Props;
-                dependencies: ContentSourceDependenciesValueFromRef<Dependencies>;
-            },
-        ],
+        [ContentSourceInput<GetRevisionProps, Dependencies>],
         Promise<ContentComputeRevisionEventResponse>,
         Context
     >;
@@ -66,12 +78,7 @@ export function createContentSource<
      * Callback to generate the document of a page.
      */
     getPageDocument: RuntimeCallback<
-        [
-            {
-                props: Props;
-                dependencies: ContentSourceDependenciesValueFromRef<Dependencies>;
-            },
-        ],
+        [ContentSourceInput<GetPageDocumentProps, Dependencies>],
         Promise<Document>,
         Context
     >;
@@ -79,26 +86,34 @@ export function createContentSource<
     return {
         sourceId: source.sourceId,
         compute: async (event, context) => {
-            const output =
-                event.type === 'content_compute_revision'
-                    ? await source.getRevision(
-                          {
-                              props: event.props as Props,
-                              dependencies:
-                                  event.dependencies as ContentSourceDependenciesValueFromRef<Dependencies>,
-                          },
-                          context,
-                      )
-                    : {
-                          document: await source.getPageDocument(
-                              {
-                                  props: event.props as Props,
-                                  dependencies:
-                                      event.dependencies as ContentSourceDependenciesValueFromRef<Dependencies>,
-                              },
-                              context,
-                          ),
-                      };
+            const output = await (async () => {
+                switch (event.type) {
+                    case 'content_compute_revision': {
+                        return source.getRevision(
+                            {
+                                props: event.props as GetRevisionProps,
+                                dependencies:
+                                    event.dependencies as ContentSourceDependenciesValueFromRef<Dependencies>,
+                            },
+                            context,
+                        );
+                    }
+                    case 'content_compute_document': {
+                        return {
+                            document: await source.getPageDocument(
+                                {
+                                    props: event.props as GetPageDocumentProps,
+                                    dependencies:
+                                        event.dependencies as ContentSourceDependenciesValueFromRef<Dependencies>,
+                                },
+                                context,
+                            ),
+                        };
+                    }
+                    default:
+                        assertNever(event);
+                }
+            })();
 
             return new Response(JSON.stringify(output), {
                 headers: {
@@ -107,4 +122,8 @@ export function createContentSource<
             });
         },
     };
+}
+
+function assertNever(value: never): never {
+    throw new Error(`Unhandled discriminated union member: ${JSON.stringify(value)}`);
 }