+++ integrations/formspree/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @gitbook/integration-formspree
 
+## 0.2.7
+
+### Patch Changes
+
+-   62c10ef: Add inputType email to formspree
+
+## 0.2.6
+
+### Patch Changes
+
+-   Updated dependencies [99ac594]
+    -   @gitbook/runtime@0.19.1
+
 ## 0.2.5
 
 ### Patch Changes