+++ integrations/openapi/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @gitbook/integration-openapi
 
+## 0.0.2
+
+### Patch Changes
+
+-   99ac594: Update OpenAPI spec parsing and take order of tags in consideration
+-   e0b1bea: Add support for openapi-operation block
+-   Updated dependencies [c24bcd1]
+-   Updated dependencies [99ac594]
+-   Updated dependencies [e0b1bea]
+    -   @gitbook/api@0.96.1
+    -   @gitbook/runtime@0.19.1
+    -   @gitbook/document@0.1.1
+
 ## 0.0.1
 
 ### Patch Changes